add feature: SITL

+++ src/acs/imu.cpp
@@ -78,7 +78,7 @@ void IMU::baroCalibration() {
 // Attention: the 'length' must be consistent with the length of the array pointed to by the 'buf'
 void IMU::getRawData(int16_t *buf, uint8_t length) {
     uint8_t stepLength = min(3, length);
-    for (size_t i = 0; i < stepLength; i++) {
+    for (uint8_t i = 0; i < stepLength; i++) {
 #if SENSOR_ACC
         *(buf + i) = accSmooth[i];
 #else
@@ -91,7 +91,7 @@ void IMU::getRawData(int16_t *buf, uint8_t length) {
 
     length -= stepLength;
     stepLength = min(3, length);
-    for (size_t i = 0; i < stepLength; i++) {
+    for (uint8_t i = 0; i < stepLength; i++) {
 #if SENSOR_GYRO
         *(buf + 3 + i) = gyroWeighted[i];
 #else
@@ -109,7 +109,7 @@ void IMU::getRawData(int16_t *buf, uint8_t length) {
 #if SENSOR_MAG
     mag->GetData(buf + 6, stepLength);
 #else
-    for (size_t i = 0; i < stepLength; i++) {
+    for (uint8_t i = 0; i < stepLength; i++) {
         *(buf + 6 + i) = 0;
     }
 #endif
@@ -117,7 +117,7 @@ void IMU::getRawData(int16_t *buf, uint8_t length) {
 
 void IMU::getAccData(int16_t *buf, uint8_t length) {
     length = min(length, 3);
-    for (size_t i = 0; i < length; i++) {
+    for (uint8_t i = 0; i < length; i++) {
 #if SENSOR_ACC
         *(buf + i) = accSmooth[i];
 #else
@@ -128,7 +128,7 @@ void IMU::getAccData(int16_t *buf, uint8_t length) {
 
 void IMU::getGyroData(int16_t *buf, uint8_t length) {
     length = min(length, 3);
-    for (size_t i = 0; i < length; i++) {
+    for (uint8_t i = 0; i < length; i++) {
 #if SENSOR_GYRO
         *(buf + i) = gyroWeighted[i];
 #else
@@ -142,7 +142,7 @@ void IMU::getMagData(int16_t *buf, uint8_t length) {
     mag->GetData(buf, length);
 #else
     length = min(length, 3);
-    for (size_t i = 0; i < length; i++) {
+    for (uint8_t i = 0; i < length; i++) {
         *(buf + i) = 0;
     }
 #endif
@@ -186,7 +186,7 @@ void IMU::updateGyro(uint32_t currentTime) {
     int16_t currentGyro[3];
     gyro->update(currentTime);
     gyro->getData(currentGyro, 3);
-    for (size_t i = 0; i < 3; i++) {
+    for (uint8_t i = 0; i < 3; i++) {
         gyroWeight[i] = currentGyro[i] + gyroWeight[i];
         gyroWeighted[i] = (gyroWeight[i] + gyroPrevWeight[i]) / 3;
         gyroPrevWeight[i] = gyroWeight[i] >> 1;
