+++ src/sensors/gyroscope.cpp
@@ -22,7 +22,7 @@ void Gyroscope::update(uint32_t currentTime)
 void Gyroscope::calibration()
 {
     calibrateSteps = 512;
-    for (size_t i = 0; i < 3; i++)
+    for (uint8_t i = 0; i < 3; i++)
     {
         offsetTotal[i] = 0;
     }
